# Optimizations and more convenient functions