This pull request adds a range of new and idiomatic functions:

    func DecodeFrom(w io.ByteWriter, src io.Reader) error
    func DecodeTo(w io.ByteWriter, bottom string) error
    func DecodedLen(bottom string) int
    func EncodeTo(out io.StringWriter, s string) (int, error)
    func EncodedLen(s string) int

In addition to these additions (whoops), the overall implementation is also a lot more optimized. The new benchmarks for these changes yield

```
BenchmarkEncode-8         798862          1491 ns/op
BenchmarkDecode-8   	  379851	      3387 ns/op
```

For reference, before these changes, they were

```
BenchmarkEncode-8          91110         11410 ns/op
BenchmarkDecode-8         115258         10075 ns/op
```