@K1ngjulien 2021-02-10
Cool! Wouldn't it be possible to implement `io.Reader` and `io.Writer` directly to be even more ideomatic? see #2