@nihaals 2021-02-10
It could probably be faster by setting `valueCharacters` at compile time and use every possible byte so the while loop (below) isn't needed at run time
https://github.com/bottom-software-foundation/bottom-go/blob/a7f861a06f2b0abd02bd406a487d71cd42342957/bottom/bottom.go#L123-L126