@diamondburned 2021-02-10
> Cool! Wouldn't it be possible to implement `io.Reader` and `io.Writer` directly to be even more ideomatic? see #2

I agree, but I think using `io.WriteStream` is more truthful to the implementation. The caller can wrap the `Writer` to add a converter, if needed.