@diamondburned 2021-02-10
The new optimizations reduced the `Decode` time to `3387 ns/op` from roughly `5000 ns/op`.