@diamondburned 2021-02-10
I have managed to squeeze even more by pre-calculating bytes combinations:

```
BenchmarkEncode-8       	 1042809	      1049 ns/op	  11.44 MB/s
BenchmarkEncodeTo-8     	  224506	      5169 ns/op	  63.26 MB/s
BenchmarkDecode-8       	  244731	      5315 ns/op	  61.52 MB/s
BenchmarkDecodeFrom-8   	  180488	      7002 ns/op	  46.70 MB/s
```