@diamondburned 2021-02-11
With the latest unsafe tweaks, I have managed to get throughput roughly 2x higher than the Rust implementation:

```
BenchmarkEncode-8       	 4635616	       246 ns/op	  48.77 MB/s
BenchmarkEncodeTo-8     	  734758	      1743 ns/op	 187.62 MB/s
BenchmarkDecode-8       	  455643	      2688 ns/op	 121.64 MB/s
BenchmarkDecodeFrom-8   	  334519	      3027 ns/op	 108.04 MB/s
```

For reference, I get `28.75 MB/s` with the naive `string(bytes)` conversion.