@diamondburned 2021-02-11
Final results:

```
BenchmarkEncode-8       	 4448227	       244 ns/op	  49.09 MB/s
BenchmarkEncodeTo-8     	  601951	      1752 ns/op	 186.64 MB/s
BenchmarkDecode-8       	  483316	      2494 ns/op	 131.10 MB/s
BenchmarkDecodeFrom-8   	  375997	      3012 ns/op	 108.58 MB/s
```