@diamondburned 2021-02-11
I think it would be nice to leave benchmark results (preferably from the CI) in the (non-existent) README.