@nihaals 2021-02-11
> I think it would be nice to leave benchmark results (preferably from the CI) in the (non-existent) README.

Agreed. I'll probably manually merge and add it in one go after I try it