@diamondburned 2021-02-11
> Only comment is that it doesn't close #2 as that requires changing the flags to support files while this only supports a string input

DecodeTo supports reading from a file, and EncodeTo does as well with a wrapper type or bufio.