@diamondburned 2021-02-11
> > DecodeTo supports reading from a file, and EncodeTo does as well with a wrapper type or bufio.
> 
> The issue is for a CLI update too. It's probably unnecessary to bundle it all into one PR though so I think having the CLI update with the glue code can be done separately

That is a fair point; I wanted to change the CLI to do that as well. Oh well.