@diamondburned 2021-02-11
> I won't push a release if you want to PR it first

I have made #6.